TS: Table/Card + View/Grid
