--- src/components/Table/Card/index.js
@@ -1,9 +1,0 @@
-// @flow
-import T from 'prop-types';
-
-const Card = ({ ..._ /*: Array<any> */ }) => null;
-Card.propTypes = {
-  children: T.func.isRequired,
-};
-
-export default Card;
