+++ src/components/Table/Card/index.tsx
@@ -1,0 +1,9 @@
+import { CardRenderer } from '../views/Grid';
+
+const Card = <RowData extends object>({
+  children: _,
+}: {
+  children: CardRenderer<RowData>;
+}) => null;
+
+export default Card;
